Added: Boilerplate for supporting various implementations of split_colours
