+++ projects/dxt-lossless-transform-common/benches/split_color_endpoints/avx2.rs
@@ -1,13 +1,13 @@
 #![allow(dead_code)]
 #![allow(unused_variables)]
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_common::transforms::split_color_endpoints::portable::portable_32;
+use dxt_lossless_transform_common::transforms::split_color_endpoints::portable32::u32;
 use safe_allocator_api::RawAlloc;
 
 // Placeholder for future AVX2 implementation
 fn bench_avx2(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        portable_32(
+        u32(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -18,7 +18,7 @@ fn bench_avx2(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAllo
 // Placeholder for future unrolled AVX2 implementation
 fn bench_avx2_unroll_2(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        portable_32(
+        u32(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -29,7 +29,7 @@ fn bench_avx2_unroll_2(b: &mut criterion::Bencher, input: &RawAlloc, output: &mu
 // Placeholder for future unrolled AVX2 implementation
 fn bench_avx2_unroll_4(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        portable_32(
+        u32(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -40,7 +40,7 @@ fn bench_avx2_unroll_4(b: &mut criterion::Bencher, input: &RawAlloc, output: &mu
 // Placeholder for future unrolled AVX2 implementation
 fn bench_avx2_unroll_8(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        portable_32(
+        u32(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
