+++ projects/dxt-lossless-transform-common/benches/split_color_endpoints/portable32.rs
@@ -1,12 +1,12 @@
 #![allow(dead_code)]
 #![allow(unused_variables)]
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_common::transforms::split_color_endpoints::portable::portable_32;
+use dxt_lossless_transform_common::transforms::split_color_endpoints::portable32::u32;
 use safe_allocator_api::RawAlloc;
 
 fn bench_portable32(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        portable_32(
+        u32(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -17,7 +17,7 @@ fn bench_portable32(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut R
 // For future unrolled implementations
 fn bench_portable32_unroll_2(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        portable_32(
+        u32(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -28,7 +28,7 @@ fn bench_portable32_unroll_2(b: &mut criterion::Bencher, input: &RawAlloc, outpu
 // For future unrolled implementations
 fn bench_portable32_unroll_4(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        portable_32(
+        u32(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -39,7 +39,7 @@ fn bench_portable32_unroll_4(b: &mut criterion::Bencher, input: &RawAlloc, outpu
 // For future unrolled implementations
 fn bench_portable32_unroll_8(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        portable_32(
+        u32(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
