+++ projects/dxt-lossless-transform-common/src/transforms/split_color_endpoints/avx2.rs
@@ -1,0 +1,30 @@
+/// Splits the colour endpoints using AVX2 instructions
+///
+/// # Arguments
+///
+/// * `colors` - Pointer to the input array of colors
+/// * `colors_out` - Pointer to the output array of colors
+/// * `colors_len_bytes` - Number of bytes in the input array.
+///
+/// # Safety
+///
+/// - `colors` must be valid for reads of `colors_len_bytes` bytes
+/// - `colors_out` must be valid for writes of `colors_len_bytes` bytes
+/// - `colors_len_bytes` must be a multiple of 4
+/// - Pointers should be 32-byte aligned for best performance
+/// - CPU must support AVX2 instructions
+#[target_feature(enable = "avx2")]
+pub unsafe fn shuffle_permute_unroll_2(
+    colors: *const u8,
+    colors_out: *mut u8,
+    colors_len_bytes: usize,
+) {
+    debug_assert!(
+        colors_len_bytes >= 4 && colors_len_bytes % 4 == 0,
+        "colors_len_bytes must be at least 4 and a multiple of 4"
+    );
+
+    // TODO: Implement AVX2 version of the color endpoint splitting algorithm
+    // For now, fallback to portable implementation
+    crate::transforms::split_color_endpoints::portable32::u32(colors, colors_out, colors_len_bytes);
+}
